backend/orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.10
numpy==2.1.2
openai>=1.52.0
cohere>=5.5.0
llama-index>=0.11.0
langgraph>=0.2.0
